+++ pkg/providers/iam/iam_errors.go
@@ -7,6 +7,8 @@ var (
 	ErrIAMListAllowedAPIs = errors.New("IAM list allowed APIs failed")
 	// ErrIAMGetCIDRBlock is returned when GetCIDRBlock fails
 	ErrIAMGetCIDRBlock = errors.New("IAM get CIDR block failed")
+	// ErrIAMListUsers is returned when ListUsers fails
+	ErrIAMListUsers = errors.New("IAM list users failed")
 	// ErrIAMGetUser is returned when GetUser fails
 	ErrIAMGetUser = errors.New("IAM get user failed")
 	// ErrIAMListCIDRBlocks is returned when ListCIDRBlocks fails